♻️ Refactor: 마이페이지 로직 변경
refactor: 마이페이지 로직 변경
